Configuração dos atendimentos e iniciado a implementação do envio de mensagens via e-mail.

+++ Consultorio.Domain/Entity/Paciente.cs
@@ -2,7 +2,7 @@
 {
     public class Paciente : Usuario
     {
-        public decimal? Altura { get; set; }
-        public decimal? Peso { get; set; }
+        public double? Altura { get; set; }
+        public double? Peso { get; set; }
     }
 }
